+++ NEWS.md
@@ -4,7 +4,7 @@
 -
 
 ### Fixed
--
+- ([#18717](https://github.com/rstudio/rstudio/issues/18717)): Fixed an issue where the diagnostics system reported a spurious parse error for indexed numeric literals, e.g. `1[TRUE]`
 
 ### Dependencies
 -